refactor: reorganize chakra model
